Add support for Value objects being implicit dependencies
As part of consolidating outside dependencies, the code I work on takes on
Value objects as implicit dependencies. That allows us to take on a dependecy
to an entire build (e.g. our compiler) rather than 500 files from it.

This has worked fine in practice for months now, but it turns out to break
when using caching, because Node.get_contents() expects all dependencies to
have the "name" attribute. This change adds that attribute to the Value class
and a test to confirm that Node.get_contents() works now.