Fix some test fixture-handling nits.
dir-fixture() did not concatenate a source list into a path
as the testing doc says it should. fix-fixture() did not either,
and the doc didn't say so; now it does (docstring and external
testing doc).

There was a way to slip through both dir_fixtures and file_fixture
with the path invalid - if you have fixture dirs defined,
and the dir/file is not found in them, you'll come out of the
loop set to the last fixture dir and not try the current dir.
This doesn't break anything, just leads to a somewhat misleading
message if the fixture really isn't found - the traceback indicates
the fixture was not found in whatever the last passed-in fixture
dir was. Now it looks like it was not found in the source testing dir.
The message can still be improved to be more descriptive.

A couple of minor Py2 removals.

Testing doc didn't mention FIXTURE_DIRS, so this was added.
A bunch of other doc fiddling.

Signed-off-by: Mats Wichmann <mats@linux.com>