Adjust appveyor build and mingw tool paths  [travis skip]
All of the AppVeyor images we test on have msys2 installed,
the 2019 image doesn't have mingw any longer, so adjust PATH
to look for the msys2 ones instead.  Also added cygwin64.

The mingw tool has the msys paths in it, but not in the correct
form to find anything: it had \msys64 and \msys, with no \bin
suffix.  Added \bin to those two (this is mentioned in bug
or clean up the messy way in which paths are looked up, but
at least the msys paths shouldn't be useless now.

Signed-off-by: Mats Wichmann <mats@linux.com>