Fixup some doc inconsistencies.   [CI skip]
PR #3464 removed SourceSignatures, TargetSignatures but left a
dangling reference in the User Guide (referred to the existence
of the functions but did not give their names, which is how
they were missed). [Fixes #3519]

PR #3506 removed SourceCode, and accidentally removed the documentation
of the Split function as well. In the PR history is a note that it
was discovered and restored, but somehow it is still missing from
master so restoring (again) with small markup changes.

Signed-off-by: Mats Wichmann <mats@linux.com>