Clean up some manpage duplication, File, Dir
Removed some dupes in the Construction Variables section -
the JAR entries plus File, Dir.

Tweaked the File and Dir function entries, and the
File and Directory Nodes section.

The end-to-end tests for File and Dir didn't test the case of
supplying a list-of-strings (thus returning a list-of-nodes)
to those two functions, added a primitive one.

Signed-off-by: Mats Wichmann <mats@linux.com>