More manpage fiddling  [ci skip]
Add 2020 year.
Use entities in more places where such are already defined.
Change more instances of option formatting so the line break
is between </listitem> and <listitem>
Commented out options (not implemented) are not inside a
<listitem> block.
A few other markup bits and indentations.

Most of this does not show in the built manpage.

Also amplified a few descriptions.

Signed-off-by: Mats Wichmann <mats@linux.com>