Add support for VS2017 Express
Needed some new logic to accomodate that tools are not in the expected
place (if real host is amd64, Express still uses Hostx86 as the tools
dir, not Hostx64).

Also needed a hack to run vcvarsall.bat with the correct argument. We
now derive the "argument" by finding the appropriate host_target
vcvars script, which has the argument to vcvarsall embedded, which
we then run instead of calling vcvarsall directly.  This allows us to
deal with the pseudo target that Express has always used - we derive
this part of the script name by using an existing lookup table.
arm targets were also added (untested).

Added some pithy comments (oh, no, the Piths ran away!)

Signed-off-by: Mats Wichmann <mats@linux.com>