test cleanups: use harness python
In a few places, a command line was built to execute a wrapper
script written in Python, but the Python used was not the one used
to invoke the test run (which is made available by TestSCons), but
the result of running test.where_is('python').  On a system which
still has the thing named 'python' resolve to Python 2, this fails,
through no fault of scons itself.

The two fixture wrapper scripts used occasionally by the tests
used subprocess.call; this is considered "old" though not marked
as deprecated at this time. Switched to subprocess.run.  See:
https://docs.python.org/3/library/subprocess.html#older-high-level-api
One of these scripts was doing unnecessary bytes-twiddling.

Modernized the inline myswig.py script (in test/SWIG/SWIG.py).
This script was reformatted using Black along the way.

Signed-off-by: Mats Wichmann <mats@linux.com>