Add a warning if msvc config cache may be outdated.
If we didn't find cl.exe in the final check in msvc_setup_env(),
issue a different warning if config caching is enabled. If it is,
there's a decent chance we've found cl.exe in the past, meaning
the cache is probably out of date - entries are indexed by path
to the bat file+arg, whereas an msvc update may bump a version
number in the path and the old path won't be valid: a cached
path entry could look like this:

"C:\\Program Files (x86)\\Microsoft Visual Studio\\2019\\Community\\VC\\Tools\\MSVC\\14.24.28314\\bin\\HostX64\\x64",

and the .28314 could be changed in an update.

Signed-off-by: Mats Wichmann <mats@linux.com>