Fix a couple of tests
After a change to the tool itself, the update-release-info test
needed an update to match.  This doesn't seem to be sufficient yet.

The SPAWN test was updated for this PR, but broke on Windows
because apparently too much quoting for calling via subprocess.
Dropped the special-case windows quoting in the test itself.

Signed-off-by: Mats Wichmann <mats@linux.com>