A feature request
Hello, 
Would it be possible to allow for a better rendering of the dependency tree ?
I implemented this change locally, and would really appreciate if i could generate this tree for documentation purposes.
Of course these have to be supported by the terminal that scons runs in. 
To combat this it would be possible to allow customization's from the project's via environmental variables.