rm-py2: Remove "from __future__" from more places
A couple of minor reformats along the way, most prominently, in tests,
if being edited anyway, make sure the docstring most tests have is
actually the docstring (sometimes the __revision__ line came before,
which makes the string not be the docstring).

Snuck in some minor framework changes that were orphaned when another
draft PR was not needed:  this almost all docstring changes, the
functional part is using casefold instead of lower in a match func -
a slightly better approach which is now possible that Py2 compatibility
is not needed.

Signed-off-by: Mats Wichmann <mats@linux.com>