Continue manpage cleanups  [ci skip]
Eliminate some "note" wording in manpage. This is a "tone" thing,
trying to avoid so many "asides", like "Note that blah blah", by
trying to reword into a more regular flow, and occasionally dropping
the aside if it doesn't seem quite suited to the manpage.

More harmonization of docbook entity usage, cleanup
of comments around generated sections, and other fiddling.

<literallayout> around a code fragment was changed to <programlisting>
<literallayout> around output was changed to <screen>
If a <screen> section contained use input, that was marked
with the <userinput> entity.  These changes ought to also
facilitate better conversion to other document markup formats,
something which has been discussed.

In the Description section, the remaining command-line discussion
at the end was merged into the earlier wording about command-line
processing, to improve the flow a bit.

Some options entries were adjusted for appearance - indents, etc.

Mention of Python 2.7 as a requirement is dropped.

Signed-off-by: Mats Wichmann <mats@linux.com>