Pass some env vars that affect msvs tool setup.
MSCommon already makes sure a few variables are propagated before
calling the the dev env setup script (vcvarsall.bat or relative).
This change adds two more: VSCMD_DEBUG enables VsDevCmd.bat
(which is eventually called) to log the environment before and
after setup, and VSCMD_SKIP_SENDTELEMETRY, if set, inhibits
the new (VS 2019) functionality to send telemetry information.

The change is motivated by the telemetry code generating undecodable
output for certain locales.  This should allow suppressing it,
but doesn't really fix the underlying problem that we're not
handling text right on Windows (on Py2 this would not have failed,
but that was an error as well - it should have).

Signed-off-by: Mats Wichmann <mats@linux.com>