Improve manpage treatment of Variables and Configure
Various cleansup to sections Configure Contexts,
Command-Line Construction Variables and SConscript Variables.

Stylesheet-level changes place multiple function signature
entries each on their own line (affects primarily Configure
Contexts).

A couple of stylesheet files were checked in in DOS format,
changed them to the same as the rest of the codebase
(this shows lots of spurious diff).

Discovered a few minor tweaks to make to code implementing
the above, so this is not completely doc-only.

Signed-off-by: Mats Wichmann <mats@linux.com>