Fix generated tools doc intermediate file.  [ci skip]
The process to generate tools.gen writes entities for links to
construction variables used/set by each tool.  When this data is
written out using lxml's tostring() method, this is encoded, causing
&foo; to become &amp;foo; which then doesn't work in later processing
as these files are included, as they're no longer valid entity
references for substitution.

This seems really hard to fix directly, because tostring() is working
as documented by doing this, so for now - maybe forever in light of
thoughts of converting docs to a different format - just postprocess
the file to undo the damage.

A hack, but fixes #3580

Signed-off-by: Mats Wichmann <mats@linux.com>