Fix inconsistencies between RootDir attributes
The RootDir class was returning different values for path and _path as well as
different values for abspath and _abspath. This is because the underscored
versions were being set in the RootDir constructor, while the non-underscored
versions were going through the EntryProxy wrapper, which is only coded to do
a simple append of paths.

I considered trying to fix EntryProxy to detect this case but instead went with
a simpler approach where RootDir overrides the attributes that it wants to
avoid EntryProxy calls. Right now I have this as path and abspath.