Add highlighting to funcs/builders  [ci skip]
When generating documentation from SConsXML markup, add some
markup to match the Python doc style better. In pseudo-markup,
we will now get:

<bold>foo</bold>(args)
<italic>env</italic>.<bold>foo</bold>(args)

That is, the function name is highlighted, and for the
environment method, env looks like a variable (which it is),
rather than the literal name env.

The methods to create an element now take keyword args,
so they can pass them on to the etree.Element method,
which understands an attribute keyword we need for one
of these changes.  Added a method to create a SubElement,
which makes usage clearer.  To complement the *Text
methods, *Tail methods were added to set/get the .tail
attribute of a node.

This change is to the tools only, and affects the result of
doing "python bin/docs-update-generated.py".

Signed-off-by: Mats Wichmann <mats@linux.com>