Update markup on TARGETS, SOURCES, ARGUMENTS  [ci skip]
Use <varname> to mark up variables in these three areas,
also a few other variables that were marked up differently.
Sometimes that markup comes in the form of using an existing
entity (&BUILD_TARGETS; for example), which is already
suitably marked up in the entity definition.

Use <parameter> to mark up when they're in the context of
function params.

The parse_flags kwarg now includes a link to the env.ParseFlags
method, since the latter describes how the args are distributed.

Signed-off-by: Mats Wichmann <mats@linux.com>