Update generation of scons-xml entities  [ci skip]
PR #3602 introduced a new problem - the documents that include
the generated functions and builders files now don't validate.
The global function/environment method name was bolded by
using <emphasis>, but Docbook doesn't allow that inside a
<literal> element (it actually works fine with the processing
tools we use, but does fail validation).

Rework the idea: use <function> and <methodname> for the markup,
and change the way those are rendered in html (man/html.xsl and
user/html.xsl) - the PDF already rendered these in bold so no
change needed there.  Also don't wrap the whole contents of the
<term> element in <literal>, which would have left the argument
list in regular font which the function name and instance
name in monospace - an odd look.  So the argument list was
wrapped in <parameter>, since that's what they are.  Don't
bother to try to parse it down into individual args, just do
the whole chunk, less the parentheses.

Signed-off-by: Mats Wichmann <mats@linux.com>