Add keys() and values() to environment
A construction environment provides a method items() which gives
you back the k, v pairs from the internal _dict, helping an env
behave like a dictionary.  This adds the keys() and values()
methods to SubstitutionEnvironment (which is the "real" consenv)
and OverrideEnvironment to further emulate a dictionary.

These methods don't need to convert to list here, the caller can do
that itself if it doesn't want a Py3 view object.

Add tests, and tweak the existing test a bit.

Also in a couple of places in Environment.py, don't loop over a view
into a dict that is being changed (del) - loop over the key view object
from a copy instead.

Signed-off-by: Mats Wichmann <mats@linux.com>