[PR #3631] clarify Detect and WhereIs  [ci skip]
Per review comment, simplfy the way the consenv values
are referred to, and add a reference to env.WhereIs.

env.WhereIs had some issues as well, so updated:
clarified the difference between env. version and
global version, and found the Note at the end to
use SCons.Util.WhereIs was not needed, since the
global function WhereIs is exactly SCons.Util.WhereIs:

  Script/__init__.py: WhereIs = SCons.Util.WhereIs

Signed-off-by: Mats Wichmann <mats@linux.com>