Eliminate Py2-compat with_metaclass jig
SCons has used a wrapper trick because the way a class declares
it uses a metaclass differs between Py2 and Py3. Drop this usage
and list the metaclass the Py3 way.

Along the way, stop reusing the type class name as a parameter,
as it might be confusing (SCons and packaging).

Signed-off-by: Mats Wichmann <mats@linux.com>