Extend `Environment.Dump()` to select serialization format
Environment.Dump() produces pretty-printable results only, so the
usefulness of this method is limited to debugging purposes.

The existing method is extended to allow selecting a serialization
format to use via a `format` parameter. Namely, it's now possible to
serialize variables as a JSON-formatted string, which makes it possible
for automated external tools to inspect the environment more easily.