Update manpage Extending section  [ci skip]
Extending SCons: this is mainly syntax: adding/adjusting markup,
reformatting code examples, adding the word "Example:" in a
few places, moving a paragraph outside a list since it applies
to the whole list, not to the last item of the list.
A list is reformatted to a <simplelist> for a shorter layout.

Signed-off-by: Mats Wichmann <mats@linux.com>