Update manpage Configuration File section   [ci skip]
CONFIGURATION FILE REFERENCE: markup added/updated, examples
reformatted.

A chunk of text followed the included "GENERATED CONSTRUCTION VARIABLE
DESCRIPTIONS" which describes how to access consvars, this was
moved before the inclusion for better flow. A bit was added to
the final paragraph before that inclusion to remind that actual
consvars are based on which tools were actually able to initialize.

Signed-off-by: Mats Wichmann <mats@linux.com>