Update forked copytree function
SCons carries a forked copy of shutil.copytree() in order to
have it not fail if the target directory exists.  Rebase the copy
it's forked off to Python 3.7 stdlib version, plus add the mod
in the style of Python 3.8 with a kwarg and using os.makedirs' ability
to do the same ignore-exists (we can't just require the 3.8 version,
since the baseline Python version is 3.5).

Signed-off-by: Mats Wichmann <mats@linux.com>