Fix testing subdir usage
Various tests called subdir several times with the same directory,
which is pointless.  In previous code, this emitted a message
though it did not fail.  Stop doing that.

The dir_fixture() method did some convoluted things to make sure
the directories to write to exist, change this around to simplify.
The subdir() method already combines the testdir, so dir_fixture
doesn't need to.  Also handle more cleanly the case of the target
directory being an absolute path, which seems to have been intended,
but would not work.

Also clean up file_fixture() along the same principles.

Change the subdir() method to not need to be given an ordered
list of directories; instead call os.makedirs on each so
intermediate steps are made automatically.  Along the way,
the print about making a directory that already existed vanishes.

Tests which did os.path.join on directory pieces when calling
file_fixture no longer do so, since file_fixture (and dir_fixture)
do joining on an arg which is a list of paths like many other scons
functions.

The testing doc was updated to fix some wording and reflect
the above changes.

Signed-off-by: Mats Wichmann <mats@linux.com>