Fix race condition when importing typelibs on Windows (PR 2161)
On Windows, if a file #import's a type library, the compiler automatically
generates .tlh and .tli files in the build directory. The compiler does not
synchronize writes of these files so they should be considered side effects.
They unfortunately have timestamps and absolute paths in them so (1) they
are not cacheable and (2) cannot be considered as sources or they will cause
unnecessary cache misses. However, setting them as normal SCons side effects
would cause all files with the same target directory with that dependency
to no longer be executed in parallel.

My solution is to introduce the idea of a temporary side effect. This
effectively prevents all nodes with that side effect from executing at the
same time until one node has successfully run. Once one node has successfully
run, then it is no longer considered to be a side effect because it is already
generated.