Make sure get_contents() method of Nodes return 'bytes'
Dir, Alias, ActionCaller compute their contents. The get_contents()
of Dir and Alias now return a utf-8 formatted string. The content
of ActionCaller is usually the 'co_code' attribute of the code object of
the function it wraps, when this is unavailable, return the 'repr' of
the wrapped function encoded with default options (currently 'utf-8').