[PR #3748] Fix typo and add some links  [ci skip]
Cleaned up the reference to some consvars only having an effect
on init by adding MSVC_VERSION as an example.  That entry already
said so, so should be non-controversial.  That change led to
a few changes in the file defining that var - mostly turning
entity references into entity link references.

Signed-off-by: Mats Wichmann <mats@linux.com>