[PR #3752] guide: update consvar setting after review  [skip appveyor]
Review comments noted that there was perhaps more assumptions
about Python knowledge than warranted, added some things to
address that, including an intro to dictionaries.

Markup changes in environment chapter, plus more wording tweaks.

Filled out an example of of overriding when calling a builder,
and added a note on when that is not sufficient, and separate
envs are needed.

Dropped a python2 paragraph.

Signed-off-by: Mats Wichmann <mats@linux.com>