Enable environment dict.setdefault
setdefault() now works on SubstitutionEnvironment and OverrideEnvironment.
Unittests added for that.

OverrideEnvironment Dictionary() method was not accepting the one-arg
and multi-arg forms that are documented for the regular env.Dictionary().
Not detected because the testcase also didn't test that; updated.
Testcase was misnamed, duplicated test_items - not sure if that
mattered or not.

Tests did not exist for OverrideEnvironment keys() and values() methods.

Signed-off-by: Mats Wichmann <mats@linux.com>