Change erorr in case not-found Tool requested
On the premise that the previous error - with a lengthy traceback
and message "SConsEnvironmentError: No module named foo:" isn't very
communicative - the error is changed to a UserError, which limits the
traceback to just the SConstruct line, and the message is updated
and now includes the tool dirs searched.  This is assuming that putting
an unknown tool (or a tool in in incorrectly specified path) into an
SConscript is an error in the SConscript, thus UserError.

The two affected unit tests are updated to expect UserError. Along the
way, an existing technique from one of the unit tests is moved over to
the other, allowing the elimination of a bare raise which was flagged by
pylint (marked TODO in the file).  The same change is then applied to
a third unit test, unrelated but marked with the same TODO.

Signed-off-by: Mats Wichmann <mats@linux.com>