test: Pass a correct flag to DVIPDFFLAGS test
Pass '-R2' instead of '-N' as the tested flag to DVIPDFFLAGS test.
Apparently '-N' alone is not a valid flag (as of ghostscript-gpl 9.53.1)
and it causes the test to fail:

  cd . && dvipdf -N foo.dvi foo.pdf
  -N must be between 2 and 4194303
  scons: building terminated because of errors.

It is not even mentioned in the manual page (though looking
at the script, it is passed to gs).  Use '-R2' instead as an arbitrary
documented flag.