Fix a bug when handling two binaries in a row with the same typelib dependency
The temporary side effect's "sources" entry wasn't being cleared after it was
processed. This was a problem when postprocessing the temporary side effect
for the second binary afterwards because the code tried to remove the temporary
side effect from sources of the first binary, when it had already been handled.
A test is included that consistently failed without the fix.