Fix occasional test failures when running multiple jobs
The way runtest.py passes the list of fixture directories is racy because it
sets it in os.environ['FIXTURE_DIRS'] and then spawns the subprocess, counting
on Python to start the subprocess before that list is overwritten when spawning
the next directory. At least on Windows, the environment is not copied in
subprocess.run so runtest.py may overwrite the list of fixture directories
with the list for test #2 while the subprocess module is still kicking off
test #1. I was able to easily reproduce this by running the command:
`python runtest.py -j 2 test\MSVC\VSWHERE.py test\AS\ASPPFLAGS.py`
a few times in a row. However, with this fix, that command repeatedly succeeds.

To validate ths fix, I also ran that command with "--xml a.xml" and
"--xml a.xml --nopipefiles" to validate that those other executors worked
correctly.