Remove some "star imports"
Usually, it's unit tests that do this, and it's not really crucial
to kill those off, but checkers do complain, including sider if you
touch anything in one of those files.

In SCons/Environment,py, removed all SCons.Util prefixes, than turned the
import of Util into a star import, running a tool on that then changes
it to import only the used symbols from Util. Since there are lots,
that ought to be a small performance win, since it doesn't have to do
namespace lookups on SCons.Util.

Signed-off-by: Mats Wichmann <mats@linux.com>