Fix invalid cache state when using SCons interactive mode
SCons.Node.Node.clear() had code to clear cache-related state but it was only
clearing a non-unused variable _calculated_sig. Fix this by clearing the three
variables now used in practice by get_cachedir_bsig() and the functions that it
calls: `cachedir_csig`, `cachesig`, and `contentsig`. Also reset `cached` to 0
because the task scheduler depends on it but doesn't reset it.