[WIP] Proof-of-concept remote caching support
This change introduces remote caching support to SCons. Currently SCons has
support for caching built objects in directories, but remote caching adds
support for sharing built objects across multiple machines. For example at
VMware, we have our official builds push to cache and all developers will
fetch from cache.

The end result of this is that developers are able to only build what they
have changed locally. All other cacheable build actions will be fetched
from the remote cache server, dramatically speeding up builds.

One piece that I introduced as part of this change is a new scheduler as a new
class ParallelV2 that is used if remote caching is enabled or if a developer
passes --use-scheduler-v2. Whereas the existing parallel scheduler waits on
jobs if the job queue is full, this new scheduler will attempt to continue
scanning for new jobs whenever possible. In addition, it supports remote
caching by draining the "pending remote cache fetches" queue just like it
drains the "pending jobs" queue. I believe that this scheduler is an
alternative to https://github.com/SCons/scons/pull/3386 that should have the
same effect in keeping jobs as high as possible. I introduced it as a separate
class for now, but it could replace the Parallel class if we wanted.

The new RemoteCache class owns the job of fetch from and pushing to a Bazel
remote cache server or any other similar server that supports /ac/ and /cas/
GET and PUT requests using SHA-256 file names. See
https://github.com/buchgr/bazel-remote for more details on the server.
This class uses urllib3 for network requests. I chose it because it has good
support for concurrency across threads using its ConnectionPool class.

As part of implementing this functionality, the following new parameters are
introduced:
	--remote-cache-fetch-enabled: Enables fetch of build output from the server
	--remote-cache-push-enabled: Enables push of build output to the server
	--remote-cache-url: Required if fetch or push is enabled
	--remote-cache-connections: Connection count (defaults to 100)