[PR #3836] add example of SConsign(dbm_modue=...)  (skip appveyor)
It wasn't clear that you need to pass the module name to
SConsignFile, the expectation might be to pass a string,
so clarified and an example added to show it.

Some of the wording was simplified a bit.

Added a testcase of explicity passing dbm_module=SCons.dblite,
as this should work.

Signed-off-by: Mats Wichmann <mats@linux.com>