Build up manpage information on sconscript files.
Tweaked introductory section with a few more details.

Added a new subsection to begin the Confguration File Reference
section, called Configuration Files, mainly as a better home
for the information on alternate SConstruct names, which seemed
a distraction at the top of the file, as well as to introduce the
concept that sconscript files are independent of each other,
which does not appear to be explicitly stated (although you
could gather the information by reading some Function entries).

Also provides quick links to other Config File subsections -
manpage is hard to navigate as it's so long.

Signed-off-by: Mats Wichmann <mats@linux.com>