[PR #3847] Make tests pass again on Travis  [skip appveyor]
With the Travis build switched to install the ldc compiler through
a snap, scons needs to be able to find a binary in /snap/bin too.

On 10.x gdc compilers (part of the GNU Compiler Collection),
the erorr message for an "unrecognized command line option"
changed to "unrecognized command-line option". Wilcarded the
changed character in the expectations of the one test which failed.

Updated docs to indicate /opt/bin is part of env['PATH'] - it
was not included, and added /snap/bin to accomodate this change.

Signed-off-by: Mats Wichmann <mats@linux.com>