Update documentation for env.Append family  [ci skip]
Try to be more descriptive of what happens in some special cases.
CPPDEFINES, LIBS, and the *PATH variables are mentioned.

Don't duplicate descriptions - AppendUnique, Prepend, PrependUnique
all point back to Append for the body of the description - the
idea is to avoid divergence. The same should happen for the
implementation, but that's a different PR.

Reformat examples with Black to continue getting example style
self-consistent within the docs (lots to go here!)

Signed-off-by: Mats Wichmann <mats@linux.com>