If calling DefaultEnviornment, do not call it env  [skip appveyor] [skip travis]
Minor niggle: env = Environment() is so ingrained, use a different
variable name in the rare places where docs save the result of
calling DefaultEnvironment() - denv.  Most instances are
internal fiddles to add the cwd to PATH for the example generator
to work, only two of these actually appear in the visible docs.

Signed-off-by: Mats Wichmann <mats@linux.com>