[PR #3877] further tweak DefaultEnvironment examples  [skip appveyor] [skip travis]
The user guide instances of fiddling ['ENV']['PATH'] in
DefaultEnviroment are further updated: they use Dir('.') to
add the cwd instead of os.getcwd().  Note all but one of these
fiddles were only to make the example generation stuff find
a dummy command that was listed in an action but normally
wouldn't exist, as it's a placeholder. Those instances are
marked with an xml comment now.

Signed-off-by: Mats Wichmann <mats@linux.com>