Correct builder function signature check in Subst
The change introduced in 52773da0faddc808056ffec34ed0112e2b42b5f7 to
check builder function signatures without relying on TypeError was both
too specific and not specific enough and can lead to confusing results.

Firstly, the check only verifies that the arguments are present and does
not verify that the arguments are in the correct order. Therefore, it is
still possible for someone to write a builder function with an incorrect
signature, having the arguments in the wrong order, that executes
successfully but has incorrect results.

Secondly, the strictness of the check - allowing *only* those arguments
in the documentation - can lead to a well-intentioned person writing
what they think is a builder function with a few extra arguments
(possibly filled by using `partial`) that then does not execute as a
builder function and has incorrect arguments passed to it.

In both of the above situations, users can become confused as to why
their builder functions are not executing properly, and especially in
the second case they can be led to believe that the functions *should*
execute properly because the documentation does not specify that *only*
the given arguments are allowed. In any case, there is no need to be so
strict - it's possible to determine the user's intent while at the same
time minimizing the likelihood of hard-to-diagnose errors.

This change makes the builder function signature check less strict in
the sense that it allows more than the required arguments, but more
strict because, unlike the previous change, those arguments must be in
the correct order. This will allow users to write partial functions that
fill in any remaining required or defaulted arguments before Subst is
called. An associated documentation change makes explicit what is
required and how to produce it, as well as what happens when the
requirement is not met.