Make target selection a subsection in manpage  [skip travis] [skip appveyor]
The selection of targets to build is important, and referred to
often. Seems it would be helpful to have it as its own section, with
an xml id tag so it can be referenced via link.  To make that happen,
it's moved to the end of introduction, otherwise have to also make
subsections for other stuff that followed it, or it would look like
those parts belonged to Target Selection.

To make it easier to read the target selection section, id and link
references are added to options, so the options described can be followed
via a click.

These are both introductory bits of work: nothing links to the new
subsection yet, and only some of the references to command line options
are linked to the new option anchors.

A note added to -c and -n to further clarify only builder targets are
affected, and to -n to clarify that the determination is necessarily
imperfect.

The --hash-chunksize option is documented now (by taking the text
formerly in md5-chunksize and updating slightly), and the work related
to the change to hash options is marked with a version-specific change
notice (4.2).

Edits are made to clarify the behavior of Default to address issue #2839,
which is that any alias used should already be defined.  Added to this
PR because it gave the opportunity to add a reference to the newly-named
section (not a hyperlink, since scons-function definitions also appear
in the User Guide).

Closes #2839.

Signed-off-by: Mats Wichmann <mats@linux.com>