Drop dictionary has_key references: Py2-ism.
In some cases, added a __contains__ method instead,
not because it necessarily was needed, but for completeness.

Also one completely unrelated change because it happened to
be sitting modified in the tree when I committed modified files:
be a little more cautious about building CHECK_METHODS in
our subclassing of the optparse Option class... current
cpython starts it at None, then fills it in, so it shouldn't
be None when we subclass.

Signed-off-by: Mats Wichmann <mats@linux.com>