Change site-dir arg handling
chat discussion: this should work to not use a site_dir:

$ export SCONSFLAGS=--site-dir=foo
$ scons --no-site-dir

Commit changes the two options to write to the same variable,
so "last one on command line wins" works out. Added a test for
this to test/site_scons/site-dir.py.

Manpage updated to clarify the order of considering SCONSFLAGS,
since order does matter.

Signed-off-by: Mats Wichmann <mats@linux.com>