Add a __iadd__ method to CLVar to support inplace adds
Added a test for in-place adding of a string containing multiple words.

The CLVar testcase in UtilTests is converted to use multicharacter
argument words to tickle the parent class's tendency to do "list()"
directly on an argument - overriding that behavior, but only for
strings, is the whole purpose of the class. That is,
"bbb" should not become ['b', 'b', 'b'] but instead ["bbb"].
We wouldn't see if that was happening when all the words in the
test were single-character.

Fixes #2399

Signed-off-by: Mats Wichmann <mats@linux.com>