Test tweaks to help testing on mingw-only win32 systems
On a system where SCons is started from the mingw bash shell, and
mingw Python is the interpreter, the secondary self-identification differs.
Use that in a few tests which otherwise make the wrong decision.

EnvironmentTests used an invalid value for CCFLAGS, and empty
string is a better choice.

Change a couple of cases where Python code was run directly so
it's prefixed by the-Python-in-use, which fails if there is not
a Windows-registered program for .py files - we didn't really want
to use the Windows-native Python anyway.

Signed-off-by: Mats Wichmann <mats@linux.com>