Test harness add a from_fw to skip calls
Main functional change is a new kwarg to skip_test() to allow calls
from inside the framework to skip an additional line of traceback in
the skip output - i.e. don't just skip the entry for skip_test, but also
the function in the fw that called it.

Other functional change is for the try block in skip_if_not_msvc()
(which is one of the internal callers of skip_test() mentioned for
the other change) to catch Exception, thus avoiding system-existing
exceptions that were caught by the existing bare except, which caused
the skip to not actually skip.

The remainder of the patch is docstring reformatting, some minor
code reformats, top-of-file license blocks, etc.

Signed-off-by: Mats Wichmann <mats@linux.com>