Avoid exception if tool set to empty
When the reproducer in issue #1742 was run on git head, it failed in a
different way than in the issue: an exception "list index out of range",
caused by the test setting CC="" in the Environment call().  While this
is a rather unuseful thing to do in general, the resulting call to
env.WhereIs should not throw an exception because the general WhereIs
function ends up indexing into something that can't be indexed. Avoid
this by returning None immediately if the list of names to look for
is empty.  Note this does _not_ fix issue 1742, it just avoids the new
problem it was failing on.

Signed-off-by: Mats Wichmann <mats@linux.com>