Fix checker-spotted issues in Environment apply_tools
If tools is a list, it is looped over, calling Tool on each;
that call overwrote the loop variable - since Tool() doesn't
return any value anyway just eliminate the assignment.

EnvironmentTests had some asserts comparing with self, which
is pointless.  Tried to determine what this should be to be meaningful.

Also quieted one test which got a default action string,
unit tests normally just print a row of dots and here it was
putting out text which interrupted the flow of dots.

Signed-off-by: Mats Wichmann <mats@linux.com>