Fiddle the gettext tool family imports/warnings more
Previous few rounds of PR #3942 all seemed to fall in some hole.
If it works right when the tool is found, it doens't work when
the tool is not found (with the main complainer being the
test/import.py test, which tries to import *everything* in
the tools dir). This change should let the tests pass on
both Posix and Windows hosts.

Signed-off-by: Mats Wichmann <mats@linux.com>