Install tool fix scons_copytree recursion
When called recursively the dirs_exist_ok parameter is passed in the wrong
position. This may modify the value of ignore_dangling_symlinks and causes
dirs_exist_ok to take on the default value of False.

The install will 'silently' fail if the destination directory exists when this
happens, and there can obviously be unintended side effects if
ignore_dangling_symlinks is changed.

The fix simply fixes the arguments in the recursive call, and also uses keyword
arguments.