Update variables/cli chapter of user guide  [ci skip]
This is really just tweaking, no material changes: the subject of
instructions is now consistently "you", and the frequent distinction
between the sconscript writer and a seperate category of "users"
is emilinated for a cleaner read, and it doesn't say "we" any longer
(this distinction is sure to still exist elsewhere in the User Guide).
Present tense is also used - there were previously a lot of future
tense uses ("will foo").

As usual, some markup stuff is added.

doc/scons.mod is condensed a little by making five-line comment
blocks into one line, and markup is changed on the three
environment variables (<envar> instead of <varname>).

Signed-off-by: Mats Wichmann <mats@linux.com>